+++ apps/paste-service/core/handler.ts
@@ -1,4 +1,5 @@
 import type { PasteStore } from "./storage";
+import { corsHeaders } from "./cors";
 
 export interface PasteOptions {
   maxSize: number;
@@ -9,6 +10,8 @@ const DEFAULT_OPTIONS: PasteOptions = {
   maxSize: 524_288, // 512 KB
   ttlSeconds: 7 * 24 * 60 * 60, // 7 days
 };
+
+const ID_PATTERN = /^\/api\/paste\/([A-Za-z0-9]{6,16})$/;
 
 /**
  * Generate a short URL-safe ID (8 chars, ~48 bits of entropy).
@@ -59,4 +62,73 @@ export class PasteError extends Error {
   ) {
     super(message);
   }
+}
+
+/**
+ * Shared HTTP request handler for the paste service.
+ * Both Bun and Cloudflare targets delegate to this after wiring up their store.
+ */
+export async function handleRequest(
+  request: Request,
+  store: PasteStore,
+  cors: Record<string, string>,
+  options?: Partial<PasteOptions>
+): Promise<Response> {
+  const url = new URL(request.url);
+
+  if (request.method === "OPTIONS") {
+    return new Response(null, { status: 204, headers: cors });
+  }
+
+  if (url.pathname === "/api/paste" && request.method === "POST") {
+    let body: { data?: unknown };
+    try {
+      body = (await request.json()) as { data?: unknown };
+    } catch {
+      return Response.json(
+        { error: "Invalid JSON body" },
+        { status: 400, headers: cors }
+      );
+    }
+    try {
+      const result = await createPaste(body.data as string, store, options);
+      return Response.json(result, { status: 201, headers: cors });
+    } catch (e) {
+      if (e instanceof PasteError) {
+        return Response.json(
+          { error: e.message },
+          { status: e.status, headers: cors }
+        );
+      }
+      return Response.json(
+        { error: "Failed to store paste" },
+        { status: 500, headers: cors }
+      );
+    }
+  }
+
+  const match = url.pathname.match(ID_PATTERN);
+  if (match && request.method === "GET") {
+    const data = await getPaste(match[1], store);
+    if (!data) {
+      return Response.json(
+        { error: "Paste not found or expired" },
+        { status: 404, headers: cors }
+      );
+    }
+    return Response.json(
+      { data },
+      {
+        headers: {
+          ...cors,
+          "Cache-Control": "public, max-age=3600",
+        },
+      }
+    );
+  }
+
+  return Response.json(
+    { error: "Not found. Valid paths: POST /api/paste, GET /api/paste/:id" },
+    { status: 404, headers: cors }
+  );
 }
